Allow 'databases' server config option to be set by ENV
This is quite specific to my needs, but possibly useful for others. 

My setup is PhpRedisAdmin running in Docker on ECS Fargate connnecting to Redis Elasticache. AWS Elasticache has a restricted CONFIG command so you can't get the databases.

This change allows me to set the databases in the Task Definition in ECS.